Refactor component imports to use local hooks
- Updated multiple components to import hooks from local paths instead of the centralized query file, improving modularity and maintainability.
- Adjusted imports in ChatAccordion, AutoSelectConversations, ConversationAccordion, and other components to streamline code organization.
- Enhanced clarity and consistency in the codebase by ensuring that related hooks are grouped together within their respective component directories.
