+++ echo/frontend/src/components/chat/ChatAccordion.tsx
@@ -1,10 +1,9 @@
 import { t } from "@lingui/core/macro";
 import { Trans } from "@lingui/react/macro";
 import {
-  useDeleteChatMutation,
   useProjectChats,
-  useUpdateChatMutation,
 } from "@/lib/query";
+import { useDeleteChatMutation, useUpdateChatMutation } from "./hooks";
 import {
   Accordion,
   ActionIcon,
