+++ echo/frontend/src/components/conversation/ConversationAccordion.tsx
@@ -2,14 +2,16 @@ import { t } from "@lingui/core/macro";
 import { Trans } from "@lingui/react/macro";
 import { Icons } from "@/icons";
 import {
-  useAddChatContextMutation,
   useConversationsByProjectId,
-  useDeleteChatContextMutation,
   useProjectChatContext,
-  useMoveConversationMutation,
   useInfiniteProjects,
   useProjectById,
 } from "@/lib/query";
+import {
+  useAddChatContextMutation,
+  useDeleteChatContextMutation,
+  useMoveConversationMutation,
+} from "./hooks";
 import { useAutoAnimate } from "@formkit/auto-animate/react";
 import {
   Accordion,
