+++ echo/frontend/src/components/conversation/MoveConversationButton.tsx
@@ -18,7 +18,8 @@ import { useEffect, useState } from "react";
 import { useInView } from "react-intersection-observer";
 import { IconArrowsExchange, IconSearch } from "@tabler/icons-react";
 import { FormLabel } from "@/components/form/FormLabel";
-import { useInfiniteProjects, useMoveConversationMutation } from "@/lib/query";
+import { useInfiniteProjects } from "@/lib/query";
+import { useMoveConversationMutation } from "./hooks";
 import { Trans } from "@lingui/react/macro";
 import { useI18nNavigate } from "@/hooks/useI18nNavigate";
 import { useParams } from "react-router-dom";
