+++ echo/frontend/src/components/project/ProjectSidebar.tsx
@@ -1,7 +1,8 @@
 import { t } from "@lingui/core/macro";
 import { Trans } from "@lingui/react/macro";
 import { Icons } from "@/icons";
-import { useCreateChatMutation, useProjectById } from "@/lib/query";
+import { useProjectById } from "@/lib/query";
+import { useCreateChatMutation } from "./hooks";
 import {
   ActionIcon,
   Box,
