+++ echo/frontend/src/components/report/UpdateReportModalButton.tsx
@@ -9,11 +9,7 @@ import {
 } from "@mantine/core";
 
 import { useEffect, useState } from "react";
-import {
-  useCreateProjectReportMutation,
-  useDoesProjectReportNeedUpdate,
-  useProjectReport,
-} from "@/lib/query";
+import { useProjectReport, useCreateProjectReportMutation, useDoesProjectReportNeedUpdate} from "./hooks";
 import { useParams } from "react-router-dom";
 import { t } from "@lingui/core/macro";
 import { languageOptionsByIso639_1 } from "../language/LanguagePicker";
