+++ echo/frontend/src/components/resource/hooks/index.ts
@@ -1,0 +1,35 @@
+import { deleteResourceById, updateResourceById } from "@/lib/api";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { toast } from "@/components/common/Toaster";
+
+export const useUpdateResourceByIdMutation = () => {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: updateResourceById,
+    onSuccess: (_values, variables) => {
+      queryClient.invalidateQueries({
+        queryKey: ["resources", variables.id],
+      });
+      queryClient.invalidateQueries({
+        queryKey: ["projects"],
+      });
+      toast.success("Resource updated successfully");
+    },
+  });
+};
+
+export const useDeleteResourceByIdMutation = () => {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: deleteResourceById,
+    onSuccess: () => {
+      queryClient.invalidateQueries({
+        queryKey: ["projects"],
+      });
+      queryClient.invalidateQueries({
+        queryKey: ["resources"],
+      });
+      toast.success("Resource deleted successfully");
+    },
+  });
+};
