+++ echo/frontend/src/routes/participant/ParticipantReport.tsx
@@ -1,10 +1,10 @@
 import { useParams, useSearchParams } from "react-router-dom";
 import { ReportRenderer } from "@/components/report/ReportRenderer";
 import {
-  useCreateProjectReportMetricOncePerDayMutation,
   useLatestProjectReport,
   useProjectReportViews,
 } from "@/lib/query";
+import { useCreateProjectReportMetricOncePerDayMutation } from "@/components/participant/hooks";
 import { LoadingOverlay, Text, Title, Stack } from "@mantine/core";
 import { Trans } from "@lingui/react/macro";
 import { Logo } from "@/components/common/Logo";
