+++ echo/frontend/src/routes/project/conversation/ProjectConversationOverview.tsx
@@ -13,9 +13,9 @@ import {
 import { useParams } from "react-router-dom";
 import {
   useConversationById,
-  useConversationChunks,
   useProjectById,
 } from "@/lib/query";
+import { useConversationChunks } from "@/components/conversation/hooks";
 import { ConversationEdit } from "@/components/conversation/ConversationEdit";
 import { ConversationDangerZone } from "@/components/conversation/ConversationDangerZone";
 import { finishConversation, generateConversationSummary } from "@/lib/api";
