+++ echo/frontend/src/routes/project/resource/ProjectResourceOverview.tsx
@@ -1,9 +1,9 @@
 import { Trans } from "@lingui/react/macro";
+import { useResourceById } from "@/lib/query";
 import {
   useDeleteResourceByIdMutation,
-  useResourceById,
   useUpdateResourceByIdMutation,
-} from "@/lib/query";
+} from "@/components/resource/hooks";
 import {
   ActionIcon,
   Box,
